+++ tests/TestCase/Command/ClearAllCommandTest.php
@@ -17,9 +17,11 @@ namespace Thumber\Cake\Test\TestCase\Command;
 
 use Cake\Console\ConsoleIo;
 use Cake\Console\Exception\StopException;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
 use Cake\Console\TestSuite\StubConsoleOutput;
 use Exception;
-use MeTools\TestSuite\CommandTestCase;
+use Thumber\Cake\Command\ClearAllCommand;
+use Thumber\Cake\TestSuite\TestCase;
 use Thumber\Cake\TestSuite\TestTrait;
 use Thumber\Cake\Utility\ThumbManager;
 
@@ -27,8 +29,9 @@ use Thumber\Cake\Utility\ThumbManager;
  * ClearAllCommandTest class
  * @property \Thumber\Cake\Command\ClearAllCommand $Command
  */
-class ClearAllCommandTest extends CommandTestCase
+class ClearAllCommandTest extends TestCase
 {
+    use ConsoleIntegrationTestTrait;
     use TestTrait;
 
     /**
@@ -37,6 +40,8 @@ class ClearAllCommandTest extends CommandTestCase
      */
     public function testExecute(): void
     {
+        $this->useCommandRunner();
+
         $command = 'thumber.clear_all -v';
 
         $this->createSomeThumbs();
@@ -48,11 +53,20 @@ class ClearAllCommandTest extends CommandTestCase
         $this->exec($command);
         $this->assertExitSuccess();
         $this->assertOutputContains('Thumbnails deleted: 0');
+    }
 
-        //On failure
+    /**
+     * Test for `execute()` method, on failure
+     * @test
+     * @uses \Thumber\Cake\Command\ClearAllCommand::execute()
+     */
+    public function testExecuteOnFailure(): void
+    {
         $this->expectException(StopException::class);
-        $this->Command->ThumbManager = $this->createPartialMock(ThumbManager::class, ['_clear']);
-        $this->Command->ThumbManager->method('_clear')->willThrowException(new Exception());
-        $this->Command->run([], new ConsoleIo(null, new StubConsoleOutput()));
+        $ThumbManager = $this->createPartialMock(ThumbManager::class, ['_clear']);
+        $ThumbManager->method('_clear')->willThrowException(new Exception());
+        $Command = $this->createPartialMock(ClearAllCommand::class, ['getThumbManager']);
+        $Command->method('getThumbManager')->willReturn($ThumbManager);
+        $Command->run(['noExisting'], new ConsoleIo(null, new StubConsoleOutput()));
     }
 }